+++ src/app/admin-transaction-history/components/TransactionDetailPopUp.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import {
   Book,
   BookTagType,
