+++ src/context/transactionAdminContext.ts
@@ -1,3 +1,4 @@
+"use client";
 import { createContext, useContext } from "react";
 
 import { Transaction } from "@/data/dto/transaction.dto";
